feat(server): Resolve absolute paths
