+++ __mocks__/path.js
@@ -1,3 +1,4 @@
 module.exports = {
-  resolve: jest.fn((...values) => values.join("+"))
+  resolve: jest.fn((...values) => values.join("+")),
+  isAbsolute: jest.fn()
 };
