+++ packages/webpack-ssr-dev-server/lib/helpers/__mocks__/paths.js
