+++ packages/webpack-ssr-dev-server/lib/helpers/__mocks__/resolveApp.js
@@ -1,0 +1,1 @@
+module.exports = jest.fn(value => `TEST_DIR+${value}`);
